backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
onnxruntime>=1.17.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
supabase>=2.4.0
httpx>=0.27.0
